Implemented Camera Selection Menu
The cameras are scanned during each render cycle. Therefore they are
only updated one a render is triggered, as a roundtrip through the wasm
code needs to be made to determine the list of cameras.

If a camera is removed it will still be shown until the next render is
issued. However it will automatically be switched to an existing camera
for the new render cycle.

If only one camera is available the camera menu stays hidden.