Add hash of file to loaded js and css file for cache busting
I realized during development that the cache seems to not be properly
busted on new releases of the repl. Therefore the build filenames now
contain a hash of their content.

Unfortunately this wasn't automatically possible for the `*.worker.js`
files. However as those (only one at the moment) should not frequently
change in the future, it shouldn't be a big problem at the moment.